+++ src/main/kotlin/org/jetbrains/intellij/platform/gradle/Constants.kt
@@ -147,6 +147,7 @@ object Constants {
             val localPluginsNormalized = Attribute.of("intellijPlatformLocalPluginsNormalized", Boolean::class.javaObjectType)
             val collected = Attribute.of("intellijPlatformCollected", Boolean::class.javaObjectType)
             val extracted = Attribute.of("intellijPlatformExtracted", Boolean::class.javaObjectType)
+            internal val pluginLibrariesFiltered = Attribute.of("intellijPlatformPluginLibrariesFiltered", Boolean::class.javaObjectType)
             val jvmEnvironment = Attribute.of("org.gradle.jvm.environment", String::class.java)
             val kotlinJPlatformType = Attribute.of("org.jetbrains.kotlin.platform.type", String::class.java)